tests: add tests
Add some tests, that verify the behavior of netifd in a real system.
The tests use a podman container as a controlled runtime environment
and to allow running of tests without root access.
netifd and its libraries are build inside of the container, to have
all dependencies available. Busybox is also built, because udhcpc6
is not available.

The tests are run by a python script, it
 - creates a network namespace
 - creates virtual ethernet pairs for simulating external devices
 - starts dhcp servers as configured in the testcase (dhcpd*.conf)
 - starts ubus
 - starts netifd
 - waits for all interfaces to come up (waitfor)
 - verifies the ip configuration, link configuration, sysctl and
   nameservers using expected-files in each testcase

The tests can be run fully automatically by running
make fetch && make container-build && make container-run-tests

or manually using
make fetch && make container-build
make container-run
-> now inside container: ./run_tests.py